Add example apps (sensor fusion, web crawler, log aggregator), v0.8.0 (#20)

+++ include/csp/part/join.h
@@ -17,7 +17,7 @@ void join(std::vector<reader<T>> inputs) {
 
     while (!inputs.empty()) {
         internal::AltMatch m;
-        internal::alt_begin(&m, chanops.data(), chanops.size(), 0);
+        internal::alt_begin(&m, chanops.data(), static_cast<int>(chanops.size()), 0);
         if (m.src && m.dst)
             *static_cast<T*>(m.dst) = std::move(*static_cast<T*>(m.src));
         internal::alt_end(&m);
